lmb-1214: fix instances and instances/index.js
