--- app/templates/custom-forms/instances.hbs
@@ -1,7 +1,0 @@
-{{page-title "Overzicht"}}
-{{breadcrumb
-  this.model.form.targetLabel
-  route="custom-forms.instances"
-  model=this.model.form.id
-}}
-{{outlet}}
