+++ app/templates/custom-forms/instances/new.hbs
@@ -1,12 +1,6 @@
 {{page-title "New"}}
-{{breadcrumb
-  "Nieuwe"
-  " "
-  this.model.form.targetLabel
-  " "
-  "instance"
-  route="custom-forms.instances.new"
-}}
+{{breadcrumb "Nieuw formulier" route="custom-forms.instances.new"}}
+
 <div class="au-c-body-container au-c-body-container--scroll">
   <div class="au-o-box">
     <ContextProvider @key="form-context" @value={{this.formContext}}>
