+++ app/templates/custom-forms/overview.hbs
@@ -52,7 +52,7 @@
     <c.body as |form|>
       <td class="au-u-1-5">
         <LinkTo
-          @route="custom-forms.instances"
+          @route="custom-forms.instances.index"
           @model={{form.id}}
           class="au-c-link au-u-medium"
         >
